LayoutTests/editing/selection/ios/selection-highlight-in-foreground-layer-expected.txt

Target
Verifies that the highlight is visible when selecting text in a composited foreground layer

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS foundSelectionHighlight is true
PASS successfullyParsed is true

TEST COMPLETE
